requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
requests
python-dotenv
